perf: split slow CLI e2es and reuse cached bundles
